+++ app.js
@@ -4,7 +4,7 @@ const cors = require("cors");
 require("dotenv").config();
 
 const contactsRouter = require("./routes/api/contacts");
-const signupRouter = require("./routes/api/signup");
+const authRouter = require('./routes/api/auth')
 const usersRouter = require("./routes/api/users");
 
 const app = express();
@@ -16,7 +16,7 @@ app.use(cors());
 app.use(express.json());
 
 app.use("/api/contacts", contactsRouter);
-app.use("/api/users", signupRouter);
+app.use("/api/auth", authRouter);
 app.use("/api/users", usersRouter);
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
