+++ controllers/auth/index.js
@@ -1,0 +1,4 @@
+const register = require("./register");
+const login = require("./login");
+
+module.exports = { register, login };
