+++ controllers/contacts/getAll.js
@@ -1,0 +1,13 @@
+const { Contact } = require('../../models');
+
+const getAll = async (req, res) => {
+  const contacts = await Contact.find({});
+  res.json({
+    status: "success",
+    code: 200,
+    data: {
+      result: contacts,
+    },
+  });
+};
+module.exports = getAll;
