+++ controllers/contacts/index.js
@@ -1,0 +1,14 @@
+const getAll = require("./getAll");
+const getContactId = require("./getContactId");
+const getPost = require("./getPost");
+const getPut = require("./getPut");
+const getDelete = require("./getDelete");
+const getPatch = require("./getPatch");
+module.exports = {
+  getAll,
+  getContactId,
+  getPost,
+  getPut,
+  getDelete,
+  getPatch,
+};
