+++ controllers/index.js
@@ -1,0 +1,8 @@
+const contacts = require("./contacts");
+const auth = require("./auth");
+const users = require("./users");
+module.exports = {
+  contacts,
+  auth,
+  users,
+};
