+++ controllers/users/getCurrent.js
@@ -1,0 +1,15 @@
+const { User } = require("../../models");
+
+const getCurrent = async (req, res) => {
+  const { email } = req.user;
+  res.json({
+    status: "success",
+    code: 200,
+    data: {
+      user: {
+        email,
+      },
+    },
+  });
+};
+module.exports = getCurrent;
