+++ controllers/users/index.js
@@ -1,0 +1,5 @@
+const getCurrent = require("./getCurrent");
+
+module.exports = {
+  getCurrent,
+};
