+++ middlewares/index.js
@@ -1,0 +1,8 @@
+const validation = require("./validation");
+const ctrlWrapper = require("./ctrlWrapper");
+const authToken = require("./authToken");
+module.exports = {
+  validation,
+  ctrlWrapper,
+  authToken,
+};
